Remove and add cart ajax and json response

+++ carts/views.py
@@ -1,11 +1,16 @@
+from django.http import JsonResponse
 from django.shortcuts import redirect
+from django.template.loader import render_to_string
 
 from carts.models import Cart
+from carts.utils import get_user_carts
 from goods.models import Product
 
 
-def cart_add(request, product_slug):
-    product = Product.objects.get(slug=product_slug)
+def cart_add(request):
+
+    product_id = request.POST.get("product_id")# Получаем id продукта из ajax запроса
+    product = Product.objects.get(id=product_id)
 
     if request.user.is_authenticated:
         carts = Cart.objects.filter(user=request.user, product=product)
@@ -17,11 +22,66 @@ def cart_add(request, product_slug):
                 cart.save()
         else:
             Cart.objects.create(user=request.user, product=product, quantity=1)
-    return redirect(request.META['HTTP_REFERER'])
+
+    user_cart = get_user_carts(request)
+    cart_items_html = render_to_string(
+        "carts/includes/included_cart.html", {"carts": user_cart}, request=request
+    )
+    response_data = {
+        "message": "Товар добавлен в корзину",
+        "cart_items_html": cart_items_html,
+    }
+
+    return JsonResponse(response_data)
 
 def cart_change(request, product_slug):
     pass
-def cart_remove(request, cart_id):
-    cart = Cart.objects.get(id=cart_id)
+
+
+def cart_remove(request):
+    if request.method != "POST":
+        return JsonResponse({"error": "Метод запроса не поддерживается"}, status=405)
+
+    cart_id = request.POST.get("cart_id")
+    if not cart_id:
+        return JsonResponse({"error": "cart_id не указан"}, status=400)
+
+    try:
+        cart = Cart.objects.get(id=cart_id)
+    except Cart.DoesNotExist:
+        return JsonResponse({"error": "Корзина не найдена"}, status=404)
+
+    quantity = cart.quantity
     cart.delete()
-    return redirect(request.META['HTTP_REFERER'])
+    user_cart = get_user_carts(request)
+
+    cart_items_html = render_to_string(
+        "carts/includes/included_cart.html", {"carts": user_cart}, request=request
+    )
+
+    response_data = {
+        "message": "Товар удален",
+        "cart_items_html": cart_items_html,
+        "quantity_deleted": quantity,
+    }
+
+    return JsonResponse(response_data)
+
+# Нерабочий вариант. Почему-то ожидает GET-запрос вместо POST запроса
+# def cart_remove(request):
+#     cart_id = request.POST.get("cart_id")
+#     cart = Cart.objects.get(id=cart_id)
+#     quantity = cart.quantity
+#     cart.delete()
+#     user_cart = get_user_carts(request)
+#
+#     cart_items_html = render_to_string(
+#         "carts/includes/included_cart.html", {"carts": user_cart}, request=request
+#     )
+#     response_data = {
+#         "message": "Товар удален",
+#         "cart_items_html": cart_items_html,
+#         "quantity_deleted": quantity,
+#     }
+#
+#     return JsonResponse(response_data)
